+++ src/celengine/solarsys.cpp
@@ -1317,7 +1317,7 @@ Vector3f SolarSystem::getCenter() const
     // TODO: This is a very simple method at the moment, but it will get
     // more complex when planets around multistar systems are supported
     // where the planets may orbit the center of mass of two stars.
-    return star->getPosition();
+    return star->getPosition().cast<float>();
 }
 
 PlanetarySystem* SolarSystem::getPlanets() const
