+++ src/celengine/universe.cpp
@@ -540,7 +540,7 @@ StarPicker::StarPicker(const Vector3f& _pickOrigin,
 
 void StarPicker::process(const Star& star, float /*unused*/, float /*unused*/)
 {
-    Vector3f relativeStarPos = star.getPosition() - pickOrigin;
+    Vector3f relativeStarPos = star.getPosition().cast<float>() - pickOrigin;
     Vector3f starDir = relativeStarPos.normalized();
 
     double sinAngle2 = 0.0;
