+++ src/celestia/gtk/dialog-star.cpp
@@ -276,7 +276,7 @@ static void addStars(sbData* sb)
         values[0] = g_strdup(ReplaceGreekLetterAbbr((stardb->getStarName(*star))).c_str());
 
         /* Calculate distance to star */
-        float d = ucPos.offsetFromLy(star->getPosition()).norm();
+        float d = ucPos.offsetFromLy(star->getPosition().cast<float>()).norm();
 
         sprintf(buf, " %.3f ", d);
         values[1] = g_strdup(buf);
